Analyze function for tabulating LS means estimates from tidied rbmi pool results.

Analyze function for tabulating LS means estimates from tidied rbmi pool results.

Usage

summarize_rbmi(
  lyt,
  ...,
  table_names = "rbmi_summary",
  .stats = NULL,
  .formats = NULL,
  .indent_mods = NULL,
  .labels = NULL
)

Arguments

lyt

(layout)
input layout where analyses will be added to.

...

additional argument.

table_names

(character)
this can be customized in case that the same vars are analyzed multiple times, to avoid warnings from rtables.

.stats

(character)
statistics to select for the table.

.formats

(named character or list)
formats for the statistics.

.indent_mods

(named integer)
indent modifiers for the labels.

.labels

(named character)
labels for the statistics (without indent).

Examples

library(rtables)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
library(broom)
library(rbmi)

data <- antidepressant_data
levels(data$THERAPY) <- c("PLACEBO", "DRUG") # This is important! The order defines the computation order later
missing_var <- "CHANGE"
vars <- list(
  id = "PATIENT",
  visit = "VISIT",
  expand_vars = c("BASVAL", "THERAPY"),
  group = "THERAPY"
)
covariates <- list(
  draws = c("BASVAL*VISIT", "THERAPY*VISIT"),
  analyse = c("BASVAL")
)
draws_vars <- set_vars(
  outcome = missing_var,
  visit = vars$visit,
  group = vars$group,
  covariates = covariates$draws
)
impute_references <- c("DRUG" = "PLACEBO", "PLACEBO" = "PLACEBO")
draws_method <- method_bayes()
analyse_fun <- ancova
analyse_fun_args <- list(
  vars = set_vars(
    outcome = missing_var,
    visit = vars$visit,
    group = vars$group,
    covariates = covariates$analyse
  )
)
pool_args <- list(
  conf.level = formals(pool)$conf.level,
  alternative = formals(pool)$alternative,
  type = formals(pool)$type
)
debug_mode <- FALSE

data <- data %>%
  dplyr::select(PATIENT, THERAPY, VISIT, BASVAL, THERAPY, CHANGE) %>%
  dplyr::mutate(dplyr::across(.cols = vars$id, ~ as.factor(.x))) %>%
  dplyr::arrange(dplyr::across(.cols = c(vars$id, vars$visit)))
data_full <- do.call(
  expand_locf,
  args = list(
    data = data,
    vars = c(vars$expand_vars, vars$group),
    group = vars$id,
    order = c(vars$id, vars$visit)
  ) %>%
    append(lapply(data[c(vars$id, vars$visit)], levels))
)

data_full <- data_full %>%
  dplyr::group_by(dplyr::across(vars$id)) %>%
  dplyr::mutate(!!vars$group := Filter(Negate(is.na), .data[[vars$group]])[1])

# there are duplicates - use first value
data_full <- data_full %>%
  dplyr::group_by(dplyr::across(c(vars$id, vars$group, vars$visit))) %>%
  dplyr::slice(1) %>%
  dplyr::ungroup()
# need to have a single ID column
data_full <- data_full %>%
  tidyr::unite("TMP_ID", dplyr::all_of(vars$id), sep = "_#_", remove = FALSE) %>%
  dplyr::mutate(TMP_ID = as.factor(TMP_ID))
draws_vars$subjid <- "TMP_ID"

data_ice <- data_full %>%
  dplyr::arrange(dplyr::across(.cols = c("TMP_ID", vars$visit))) %>%
  dplyr::filter(is.na(.data[[missing_var]])) %>%
  dplyr::group_by(TMP_ID) %>%
  dplyr::slice(1) %>%
  dplyr::ungroup() %>%
  dplyr::select(all_of(c("TMP_ID", vars$visit))) %>%
  dplyr::mutate(strategy = "MAR")

draws_obj <- draws(
  data = data_full,
  data_ice = data_ice,
  vars = draws_vars,
  method = draws_method
)
#> 
#> SAMPLING FOR MODEL 'MMRM' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.001017 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 10.17 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 1200 [  0%]  (Warmup)
#> Chain 1: Iteration:  120 / 1200 [ 10%]  (Warmup)
#> Chain 1: Iteration:  201 / 1200 [ 16%]  (Sampling)
#> Chain 1: Iteration:  320 / 1200 [ 26%]  (Sampling)
#> Chain 1: Iteration:  440 / 1200 [ 36%]  (Sampling)
#> Chain 1: Iteration:  560 / 1200 [ 46%]  (Sampling)
#> Chain 1: Iteration:  680 / 1200 [ 56%]  (Sampling)
#> Chain 1: Iteration:  800 / 1200 [ 66%]  (Sampling)
#> Chain 1: Iteration:  920 / 1200 [ 76%]  (Sampling)
#> Chain 1: Iteration: 1040 / 1200 [ 86%]  (Sampling)
#> Chain 1: Iteration: 1160 / 1200 [ 96%]  (Sampling)
#> Chain 1: Iteration: 1200 / 1200 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 1.30801 seconds (Warm-up)
#> Chain 1:                4.45582 seconds (Sampling)
#> Chain 1:                5.76384 seconds (Total)
#> Chain 1: 
#> Warning: The largest R-hat is 1.18, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#r-hat
#> Warning: The Effective Sample Size is below 40% for 2 parameters. Please consider increasing burn-in and/or burn-between, or the number of samples
impute_obj <- impute( # @TODO: add support of `update_stategy` argument
  draws_obj,
  references = impute_references
)

ref_levels <- levels(impute_obj$data$group[[1]])
names(ref_levels) <- c("ref", "alt")
analyse_fun_args$vars$subjid <- "TMP_ID"
analyse_obj <- do.call(
  analyse,
  args = list(
    imputations = impute_obj,
    fun = analyse_fun
  ) %>%
    append(analyse_fun_args)
)
pool_obj <- do.call(
  pool,
  args = list(
    results = analyse_obj
  ) %>%
    append(pool_args)
)

h_tidy_pool(pool_obj$pars[1:3])
#>   group       est    se_est lower_cl_est upper_cl_est   est_contr  se_contr
#> 1   ref -1.615820 0.4862316    -2.575771   -0.6558685          NA        NA
#> 2   alt -1.707626 0.4749573    -2.645319   -0.7699335 -0.09180645 0.6826279
#>   lower_cl_contr upper_cl_contr   p_value relative_reduc
#> 1             NA             NA        NA             NA
#> 2      -1.439497       1.255884 0.8931772     0.05681725
df <- tidy(pool_obj)
df
#>   group       est    se_est lower_cl_est upper_cl_est   est_contr  se_contr
#> 1   ref -1.615820 0.4862316    -2.575771   -0.6558685          NA        NA
#> 2   alt -1.707626 0.4749573    -2.645319   -0.7699335 -0.09180645 0.6826279
#> 3   ref -4.201801 0.6515251    -5.488575   -2.9150272          NA        NA
#> 4   alt -2.869824 0.6336333    -4.121133   -1.6185159  1.33197653 0.9173825
#> 5   ref -6.358246 0.7137948    -7.769146   -4.9473458          NA        NA
#> 6   alt -4.145472 0.6982117    -5.525641   -2.7653029  2.21277417 1.0150456
#> 7   ref -7.606447 0.8018633    -9.193576   -6.0193192          NA        NA
#> 8   alt -4.742738 0.7826948    -6.291871   -3.1936053  2.86370936 1.1491282
#>   lower_cl_contr upper_cl_contr    p_value relative_reduc visit conf_level
#> 1             NA             NA         NA             NA     4       0.95
#> 2     -1.4394968       1.255884 0.89317724     0.05681725     4       0.95
#> 3             NA             NA         NA             NA     5       0.95
#> 4     -0.4800052       3.143958 0.14851295    -0.31700134     5       0.95
#> 5             NA             NA         NA             NA     6       0.95
#> 6      0.2054082       4.220140 0.03098527    -0.34801644     6       0.95
#> 7             NA             NA         NA             NA     7       0.95
#> 8      0.5867999       5.140619 0.01416790    -0.37648447     7       0.95

afun <- make_afun(a_rbmi_lsmeans)

basic_table() %>%
  split_cols_by("group", ref_group = levels(df$group)[1]) %>%
  split_rows_by("visit", split_label = "Visit", label_pos = "topleft") %>%
  summarize_rbmi() %>%
  build_table(df)
#> Visit                                       ref                alt       
#> —————————————————————————————————————————————————————————————————————————
#> 4                                                                        
#>   Adjusted Mean (SE)                   -1.616 (0.486)     -1.708 (0.475) 
#>     95% CI                            (-2.576, -0.656)   (-2.645, -0.770)
#>   Difference in Adjusted Means (SE)                       -0.092 (0.683) 
#>     95% CI                                               (-1.439, 1.256) 
#>     Relative Reduction (%)                                     5.7%      
#>     p-value (RBMI)                                            0.8932     
#> 5                                                                        
#>   Adjusted Mean (SE)                   -4.202 (0.652)     -2.870 (0.634) 
#>     95% CI                            (-5.489, -2.915)   (-4.121, -1.619)
#>   Difference in Adjusted Means (SE)                       1.332 (0.917)  
#>     95% CI                                               (-0.480, 3.144) 
#>     Relative Reduction (%)                                    -31.7%     
#>     p-value (RBMI)                                            0.1485     
#> 6                                                                        
#>   Adjusted Mean (SE)                   -6.358 (0.714)     -4.145 (0.698) 
#>     95% CI                            (-7.769, -4.947)   (-5.526, -2.765)
#>   Difference in Adjusted Means (SE)                       2.213 (1.015)  
#>     95% CI                                                (0.205, 4.220) 
#>     Relative Reduction (%)                                    -34.8%     
#>     p-value (RBMI)                                            0.0310     
#> 7                                                                        
#>   Adjusted Mean (SE)                   -7.606 (0.802)     -4.743 (0.783) 
#>     95% CI                            (-9.194, -6.019)   (-6.292, -3.194)
#>   Difference in Adjusted Means (SE)                       2.864 (1.149)  
#>     95% CI                                                (0.587, 5.141) 
#>     Relative Reduction (%)                                    -37.6%     
#>     p-value (RBMI)                                            0.0142